requirements.txt
git+https://github.com/facebookresearch/fvcore.git
torchvideo
av
torch==2.1.0 torchvision==0.16.0 --index-url https://download.pytorch.org/whl/cpu
fastapi
uvicorn
python-multipart
albumentations